+++ Tyuiu.VolodinaAA.Sprint3.Task0.V27.Test/DataServiceTest.cs
@@ -11,7 +11,7 @@ namespace Tyuiu.VolodinaAA.Sprint3.Task0.V27.Test
         public void GetSumSeriesValid()
         {
             DataService ds = new DataService();
-            double value = 5;
+            int value = 5;
             int startValue = 1;
             int stopValue = 13;
             double S = ds.GetSumSeries(value,startValue,stopValue);
